requirements.txt
Flask
mysql-connector-python
pytest
mock
